Reimplement 'unique_misses'
According to the flamegraphs, 'unique_misses' is a rather expensive
operation, due to its use of sort with a very complex predicate. This
change attempts to optimize it. At least for the fastly benchmark we get
about 2% improvement.

In additition criterion benchmark reports an 86% speedup:

unique_misses           time:   [3.8204 us 3.8808 us 3.9464 us]
                        change: [-86.347% -86.027% -85.723%] (p = 0.00 < 0.05)
                        Performance has improved.

Change-Id: Ic6e9394ceeeffc6e15ad1d11d22a3ef6ee5fc0f5
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/272
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>