Make stmt meta an associated type on UpstreamDB
Postgresql prepared statements only provide us the *types* of the
parameters to a prepared statement, not the full column metadata - this
means that having Column be an associated type on UpstreamDatabase and
having UpstreamPrepare have `params: Vec<DB::Column>` doesn't work.
Instead, this commit refactors UpstreamDatabase to have a new
StatementMeta associated type, which is used instead of *both* `schema`
and `columns` on UpstreamPrepare. The MySqlUpstream now defines a new
struct for this associated type, with the original `schema` and
`columns` fields there. The PostgreSqlUpstream didn't originally have
column metadata being sent through, and that hasn't changed here - as of
this commit, the UpstreamMeta for PostgreSqlUpstream is just (), and
will be implemented in a subsequent commit.

Change-Id: I40363c01f1b398580baad60f844297ff31bd59fe
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/277
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>