Collapse Upstream Connectors to Single Connector
This change collapses the seperate read/write upstream and noria
connectors stored in the backend to a single connector for upstream and
noria respectively. This matters for upstream transactions and prepared
statement support. Upstream transactions must be executed over a single
connector, and may contain reads and writes. Upstream prepared
statements only exist on a single connector and sending a prepared read
or write to the corresponding connector cannot be guaranteed.

Collapsing the noria reader and writer into a single connector is not
necessary, but simplifies the adapter and provides no obvious drawbacks.
Currently, we expect most customers to run in a configuration that uses
the upstream connector for writes.

The backend now contains a noria and upstream connector. The upstream
connector is wrapped in an option. The value of this option is used to
determine whether to run fallback and execute writes on upstream or
noria.

Change-Id: Ie9a609f515bf13cd56d69d5c059e164726452cee
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/278
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>