Add build config for readyset-psql image
This reverts commit 3ec2a92595a6e8c51cc0000cdba44556ee927f3b, which was
itself a revert of commit a9575af6fafc728acd15c2e04568054bec5385f9.

Re-add build-configuration, patterned off of the current configuration
for readyset-mysql, for a new readyset-psql image. This is still running
only on the main branch, but I'll test it once without the branch
filtering to make sure it works this time

Change-Id: I6387e3c9e2d46fcce84a7429b0e78a926102bd85
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/279
Reviewed-by: Harley Klein <harley@readyset.io>
Tested-by: Buildkite CI