Add release arg support to readyset-mysql
This is exactly the same as what's happening in the readyset-server
image - in fact, when I added it there, I *intended* to add it here, but
didn't - even though buildkite is passing this arg on main builds.

The idea here, as with the other image, is speeding up builds for CLs
while still getting release builds for main.

Change-Id: Ia691a3b7b9245cbb4ed954d0e3392fd257064dc6
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/287
Tested-by: Buildkite CI
Reviewed-by: Harley Klein <harley@readyset.io>