Add transaction parsing to nom-sql
This commit adds the parsing of the transaction related queries to
nom-sql. This includes all valid variants, `START TRANSACTION` `BEGIN`
and `BEGIN WORK` for initiating a transaction. `COMMIT` and `COMMIT
WORK` for commiting a transaction. `ROLLBACK` and `ROLLBACK WORK` for
rolling back a transaction.

It does not yet include parsing of
transaction configuration parameters, which can be extensive. It also
does not include parsing `SAVE POINT`, which is a transaction feature
exclusive to postgres. These will be added in a future CL.

Change-Id: I68d342514c804e339bff25cbf15abad324e2d630
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/289
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>