Write StatusFlags back to client
We should be writing StatusFlags that we get during fallback, back to
the client. Not doing so is VERY bug prone, as many status flags are read
by ORMs and relate to transaction states.

This commit ensures that we capture status flags from the last ok packet
we receive from mysql via mysql_async, and write those back when we
write either Ok or EoF.

Change-Id: Iea0cd21ae7edb9de8b4e44a5d06e730f1e846518
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/291
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>
Reviewed-by: Michael Cronce <mc@readyset.io>
Reviewed-by: Dan Wilbanks <dan@readyset.io>