Add initial implementation of Criterion benchmarks
We want to have some micro benchmarking in place to have better
justification for performance improvements. The best framework
for that is Criterion, but it has two shortcomings:

* It can only run benchmarks from dedicated bench targets
* It can only benchmark public APIs

Therefore we implement the benchmarks for the private APIs under
the "bench" feature instead, then export the benchmark functions
and call them from the Criterion benchmark executable manually.

Change-Id: I9295a561d42dcc4b8eef73dffff4480c0f8723d0
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/292
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>