Parse DOUBLE PRECISION
Parse the DOUBLE PRECISION type, as an alias for SqlType::Double.

The aliasing itself is done directly in the parser here - there could
potentially be a case to be made for doing that as a pass later, but for
now nothing cares about the difference between DOUBLE and DOUBLE
PRECISION and we don't have a desugared HIR (meaning we'd have to panic
when pattern matching on the hypothetical SqlType::DoublePrecision
variant post-desugaring) so it's simpler to do here.

Change-Id: I7c631bebd075a1379a06323c2c228d7ab9bac6be
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/304
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>