Fixes to Docker mirror pipeline generator
Most silly mistake was forgetting to set the script to be executable.
That is fixed.

When going through the images we use I found a couple missing ones and
so added those to the list of images.

Running a test also ran into the rate limit so only doing one pull at
a time.

Change-Id: I36ea310ac8f8b4acb903e955b7b05f523ea0ba99
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/309
Reviewed-by: Michael Cronce <mc@readyset.io>
Reviewed-by: Griffin Smith <griffin@readyset.io>
Tested-by: Buildkite CI