Add structure and flag for query coverage analysis
Define a new data structure, QueryCoverageInfo, which will allow us to
record information, during the execution of a noria adapter, about the
set of queries that are being run and which we support vs which go to
fallback, plus differences in execution time, and which return different
data in upstream vs noria. This struct is wrapped in a `Box::leak`ed
Mutex, and, if a new --coverage-analysis flag is passed to any adapter,
is stored as a field of any Backends initialized for that adapter.

Fixes: ENG-551
Change-Id: I94f8451cf51b3010ff6db25622c864fb15e1ef45
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/312
Reviewed-by: Peter Farr <peter@readyset.io>
Reviewed-by: Dan Wilbanks <dan@readyset.io>
Reviewed-by: Michael Cronce <mc@readyset.io>
Tested-by: Buildkite CI