Modify LocalAuthority to support ephemeral keys.
By allowing LocalAuthority to have ephemeral keys, we can more
accurately mimic using external authorities (Zookeeper, Consul).
This is accomplished by adding a new function `write_ephemeral`,
which performs a write and stores the key in a set of keys
to be removed when the authority is dropped.

A followup CL will modify worker registration to require ephemeral
keys. Workers will each register with an ephemeral key and when
the session is ended, the ephemeral key will be removed allowing
the controller to see that a worker has failed.

Change-Id: If479224cf4bf3449d2262bddb6fb191ed5b47d5d
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/313
Tested-by: Buildkite CI
Reviewed-by: Harley Klein <harley@readyset.io>