Delegate worker registration and failure detection to the authority.
This CL introduces a new type of authority state
`AuthorityWorkerState`, which tracks the state of active
workers in the system. When changes to this state occur
updates are propagated to the controller about what workers
have failed and what workers have joined the system.

This pushes failure detection out of our own system and
delegates it to the authority. Each authority is responsible
for detecting when a worker is no longer in the system. This
is done for all existing authorities (LocalAuthority and
ZookeeperAuthority) by using ephemeral keys, keys that are
deleted when the session that creates them is closed. Each
worker registers with an ephemeral key and the current
controller repeatedly checks the path that these ephemeral
keys are registered at for new keys (new workers), and the
absence of previously seen keys (worker failures).

In our integration tests we currently do not have the ability
to completely cleanup local state, specifically the
LocalAuthority, as references to the object are spread across
multiple local components that are not cleaned up. As a result,
we cannot "close" the LocalAuthority and cleanup the ephemeral
state naturally. To temporarily circumvent this, we add a new
function, `delete_ephemeral` to the LocalAuthority API.

Change-Id: I3937fdecd0558f4d779d8b943ec3027c065f8138
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/290
Tested-by: Buildkite CI
Reviewed-by: Harley Klein <harley@readyset.io>