Update deps for mysql_async
This commit updates deps for mysql_async. Our fork of mysql_async was
recently updated so that our commits were cherry picked to be on top of
newest upstream.

This also includes removing unnecessary duplicate implementations, which
existed because of library version mismatch. It also updates logic
around the use of functions from mysql_async, whose function signatures
and correct usage has since changed.

The order for safely updating deps is that we depend upon that branch of
mysql_async in our fork (so we don't break existing builds pointed at
master) and once we've switched over, we will force push that branch to
be our new master branch of our fork, and then one more CL to update the
mono-repo to point at master of our fork rather than a dedicated branch.

Change-Id: I2f71ea982bb96c087df0c8b399541e6de94f4b23
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/314
Reviewed-by: Griffin Smith <griffin@readyset.io>
Reviewed-by: Michael Cronce <mc@readyset.io>
Tested-by: Buildkite CI