noria-client-adapter: Use readyset-logging
Swap out all uses of `slog` for `tracing` in the noria_client_adapter
crate, using the new `readyset-logging` crate to initialize logging
configuration. This commit also incidentally improves logging somewhat,
by using tracing spans in a few more places and logging a few more
fields than we were previously. On the other hand, since we're no longer
building a slog logger, we no longer have anything to pass to
`ZookeeperAuthority::log_with`, so that logging won't be happening
anymore - that'll come back in a future commit when we swap out slog for
tracing there as well.

Refs: ENG-558
Change-Id: Ibf355a6150e4c01a3a54a09adf44f47bda8e6022
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/321
Tested-by: Buildkite CI
Reviewed-by: Peter Farr <peter@readyset.io>