noria-dataflow: Swap out slog for tracing
Replace all uses of `slog` with `tracing` in the noria_dataflow crate,
including removing the `log` field on Domain and hence DomainBuilder.
This latter bit loses some extra context on the log events, since that
extra logger field had the domain's index and shard on it, but that'll
be replaced once the server crate gets slog swapped out from tracing as
well.

Refs: ENG-558
Change-Id: I820d856f9d8fa12fc25b3550087764a24085a5cd
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/325
Tested-by: Buildkite CI
Reviewed-by: Vlad Krasnov <vlad@readyset.io>