treewide: Never #[macro_use] extern crates
As of the rust 2018 edition, macros can be imported directly in the
scope they're used via a regular `use` statement, versus the previous
method of importing macros which required a `#[macro_use] extern crate
<blah>` declaration at the root of the crate (eg in a lib.rs or main.rs
or whatever). This is great for a few reasons: for one, it makes it
textually obvious in the code (eg, without using any IDE support or
whatever) where a particular macro call is coming from and what it
means. Also, it avoids any sort of global-scoping of macro name
resolution (you can use multiple macros with the same name in different
modules in the same crate) The latter is especially relevant to me as I
ramp up into replacing `slog` with `tracing` and would like to do so in
smaller chunks than entire crates, but even if that wasn't the case imo
getting rid of all our macro_use'd extern crates is a great idea for the
first reason.

This commit removes all `#[macro_use] extern crate` declarations from
all crates in the codebase *except* for noria/applications (which is not
really code that we care about right now), adds
`#![deny(macro_use_extern_crate)]` to all of those crates to prevent any
macro_uses from sneaking back in, and updates all places that were
depending on globally available macros to `use` those macros locally.
That last bit makes this a relatively large change (at least if you look
at just file count) but the actual changes that're being made are
basically entirely mechanical.

Change-Id: I71d060d0c0afb7c662de18ede1301fa70fa7f166
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/318
Tested-by: Buildkite CI
Reviewed-by: Justin Miron <justin@readyset.io>