mir: Swap slog for tracing
Replace all uses of `slog` with `tracing` in the noria-mir crate.

Note that this is a 1:1 verbatim translation - I've elected *not* to do
any actual improvements to the logs here (notably by swapping out string
formatting for structured log fields) since all the logs are trace-level
events that seem to just be for debugging.

Refs: ENG-558
Change-Id: I0ad0b99bb1b845d1d2cd815541f6d90880fab06c
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/326
Tested-by: Buildkite CI
Reviewed-by: Peter Farr <peter@readyset.io>