Better fallback for system variable queries in MySQL.
Moved the fallback logic to work with the parsed `SqlQuery` instead of manually
parsing the query string (as `griffin@readyset.io` correctly pointed out).
Extending the AST for this didn't make sense, since this particular way of
dealing with system variables (selecting columns with a `@@` prefix) it's MySQL
specific.
In this approach, we parse the query and then check if there is _any_ column
which name has a `@@` prefix.
In case we need to fallback but fallback is not enabled, we return hardcoded
values (some of them are needed by some libraries).

Change-Id: I5f8469c27f25fd5fec80a10ff1e437e288a229c7
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/317
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>