Recreate database in between fallback tests
Add an extra step to noria_client::test_helpers::setup, per adapter,
which (if fallback is enabled) first connects to a management
db (`postgres` for postgres, no database for mysql) and DROPs then
CREATEs the `noria` database before running any of the other setup
process. This provides better isolation between separate tests and in
between subsequent runs of the same tests, which might be eg CREATEing
the same tables.

Change-Id: Id21795b2c30f8a1b5715999c32833265c13ca16a
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/337
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>