Send CREATE {TABLE,VIEW} to postgres *and* noria
PostgreSQL doesn't support replication of DDL changes via the method
we're using for replicating data changes - this means that things like
eg integration or logictests that depend on being able to create tables
in upstream and have them replicated to noria don't work. There are a
number of different workarounds to this issue with a number of different
tradeoffs - the one we're picking here, mostly for expedience's sake, is
to just send schema changes made to the adapter to both the postgresql
upstream and noria. This will mean we have to instruct users not to make
any schema changes directly to the upstream db.

Since schema changes *do* go through replication for MySQL, and we don't
want to turn that off (because it works perfectly fine), this is enabled
via a (defaulting to false) flag on the noria-client Backend, which is
passed through the noria-client Adapter and set to true by noria-psql

There are a couple of new integration tests here in both noria-psql and
noria-mysql testing that CREATE TABLE still behaves as expected - in the
postgresql case this also required adding an extra sleep() call before
the initial CREATE TABLE due to a (documented in the code) race
condition with the initial replication snapshot.

Fixes: ENG-560
Change-Id: I7aae4ec31a45aded60fd102d3a7b5ddc239bf931
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/338
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>