When listing changed files for CI, check just the current commit.
As an artifact of the old branch/merge Github-style workflow, we were
comparing HEAD with origin/main.  However, since our fancy new Gerrit
style workflow makes every commit its own unit of work, each of which is
individually tested with Buildkite, that doesn't make a ton of sense.

With this change, we compare the current commit with its predecessor.

Change-Id: I9508de119f248bd0bb727b49b9ab683c90d330a2
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/339
Reviewed-by: Harley Klein <harley@readyset.io>
Tested-by: Buildkite CI