Set autocommit for sqlalchemy framework test
Setting autocommit=0 (false) is a trick to effectively always be in a
transaction. It's the same as automatically writing `START TRANSACTION`
the moment that we write `COMMIT` or `ROLLBACK`. Even if we could
support this, it would be pointless, because it would mean that
literally 100% of incoming queries would go to fallback, making ReadySet
useless.

This commit enforces that autocommit is set to true, or rather SET
autocommit=1 is issued rather than SET autocommit=0. This is how
customers should setup SqlAlchemy to interact with ReadySet, otherwise
they will never be using the cache.

Note that when this is set what SqlAlchemy actually does is issue these
two queries back-to-back:

SET autocommit=0
SET autocommit=1

This means that we can't just error out if we ever see `SET
autocommit=0`. The correct behavior should be that we error out if it's
followed by anything other than `SET autocommit=1`.

This will be covered in a future CL.

Change-Id: I929c17cac7c575422fe467c8e5e2cd884084bb72
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/343
Tested-by: Buildkite CI
Reviewed-by: Michael Cronce <mc@readyset.io>