Use docker save to export Docker image artifacts
After exploring permissions on ECR, there were a number of limitations
which would make it hard to share with customers.

Instead of installing a much more complicated Docker registry, we can
use docker save to generate an artifact which others can docker load
into their own registries and such as they desire and use S3's rich
permissions to control access.

This is the first step towards a full set of pipelines for doing this.
This also triggers another pipeline on main builds and puts the block
step in that pipeline instead of in the main pipeline.

Change-Id: I5ff84b203ac662a3fbb9274d65cb9ead9606274b
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/348
Tested-by: Buildkite CI
Reviewed-by: Michael Cronce <mc@readyset.io>
Reviewed-by: Griffin Smith <griffin@readyset.io>