Make COMMIT and ROLLBACK always passthrough
Previously I had tried to error out if we received a COMMIT or ROLLBACK
when we weren't already in a transaction. This is incorrect behavior.
MySQL will always return an Ok packet for Commit or Rollback, regardless
of current transaction status.

This corrects this behavior by always passing these through to the
upstream database.

Change-Id: I46f746afc553207072885676fcee8843f1857fe1
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/354
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>
Reviewed-by: Dan Wilbanks <dan@readyset.io>
Reviewed-by: Michael Cronce <mc@readyset.io>