Handle COM_INIT_DB correctly
Many ORMS and frameworks will send an explicit COM_INIT_DB packet, which
is equivalent to `USE <DATABASE-NAME>`. We correctly parsed this packet
and turn it into a `Command::Init` type. We then would call
`self.shim.on_init()` which still had the default implementation of
simply returning Ok(()). This meant that we never wrote back an Ok
packet to the client, causing indefinite connection hanging for most
frameworks and orms.

This commit correctly handles COM_INIT_DB packets by checking if the
database wanting to be used is already the database we're connected to
over fallback. If it's not we will write back an error to the client
because as it currently stands we can only replicate from a single
dataase. In the future we should probably change this as it's common for
client applications to change databases dynamically.

In the case that fallback isn't setup, we don't have a concept of
databases anyways, so we just write back an Ok packet.

As part of this commit I needed to change the `QueryResult` enum
variants because of an enforced clippy rule, that doesn't allow variants
to all have the same postfix. In this case, they all had the same
postfix of `Result`. You can read more about this clippy rule here:
https://rust-lang.github.io/rust-clippy/master/index.html#enum_variant_names

Change-Id: I3d5ac7451cdca1050aee7945aea89f7e35454f92
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/355
Tested-by: Buildkite CI
Reviewed-by: Michael Cronce <mc@readyset.io>
Reviewed-by: Griffin Smith <griffin@readyset.io>