Change versioning for compatibility
Previously we were hard coding that our version was based off 5.1, which
is ancient. This caused laravel (framework) to try to set sql modes so
that they match with the current defaults for 5.7 and beyond. We then
send those sql modes to fallback, which in most cases is beyond 8+, and
error out because the mode `NO_AUTO_CREATE_USER` is now removed in
8.0.11+.

Ultimately we should be getting the version from fallback directly.
Having said that, this is significantly better than what we were doing
before, and is a step in the right direction until we can dedicate more
engineering time to a better solution.

This also gets laravel framework testing to pass.

Change-Id: Id4b24ac97a6881255ba3e40208fb2174ac859b47
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/356
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>
Reviewed-by: Dan Wilbanks <dan@readyset.io>
Reviewed-by: Michael Cronce <mc@readyset.io>