Provide SOFT_FAIL env variable to allow CI to "pass" with a failing step
This is desirable to run (potentially known failing) framework tests; we
want to get the results of the test against Noria, but we don't want to
block merging a CL that adds/modifies a valid test that simply doesn't
work with our software yet.

This is necessary to that goal because we explicitly check the return
code for reporting status back to Gerrit.  (Which we have to do because
Buildkite just doesn't expose its notion of failure to hooks.)

Change-Id: I64af8fecedc4f87facecf1e41a414cf9b958e96b
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/361
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>