Fix server crashing on column not found
We currently crash if a column is not found. This was surfaced by the
sqlx framework test. This commit changes the function where this error
occurs to return a ReadySetResult. Luckily every function that calls
this already returned a ReadySetResult, so the refactor was relatively
minimal.

Confirmed that the server no longer crashes, and we effectively send the
query to fallback in the adapter.

Change-Id: Icf91b08922078c0aac4b3782649823d23c6c1a4c
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/360
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>