Don't abort on task failure in tests
Add a new noria_server Config field, `abort_on_task_failure`, which
defaults to true but if set to false will disable the abort-on-panic (or
failure) behavior introduced in df9fc69ad (Abort the process if the
important tasks fail, 2021-08-16), and then set this field to false in
all places we're initializing a noria-server in a background thread as
part of an existing process: integration tests, noria-logictest, and the
query generator benchmark harness. This prevents errors that occur in
the noria server task from causing an unrecoverable full-process abort
in these cases, which is not only difficult to debug, but also
frequently not actually an error we care about - in this particular
case, this was triggered by the cleanup process for integration
tests (specifically the Drop impl for Deployment) removing the zookeeper
node out from under a noria server while it was running, which is
expected behavior and not actually something we need to worry about.

Change-Id: I4995aeb7f7841aa005a71573d8c7dabcd0ac3dc6
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/374
Reviewed-by: Justin Miron <justin@readyset.io>
Tested-by: Buildkite CI