psql-srv: Don't send ReadyForQuery after exec error
A careful reading of
https://www.postgresql.org/docs/13/protocol-flow.html indicates that
error responses during an Execute request (called "extended query" in
the documentation) in the postgresql protocol are intended to be sent by
the server, then the server is supposed to *discard* all subsequent
messages received until the client sends a Sync message, and *only then*
respond with a ReadyForQuery message. Previously, we'd send the error
message immediately followed by a ReadyForQuery message (as part of a
Response::Message2), which would then cause the client to return an
error (in the case of tokio_postgres, an "UnexpectedMessage" error) on
the second ReadyForQuery message it would receive after sending a Sync.
This begins the process of fully correctly supporting this flow by
replacing the "is_starting_up" boolean in Protocol with a new State
enum, which gets variants for StartingUp and Ready, but also
Extended (where we don't send ReadyForQuery with an error response) and
Error (currently unused, but the idea is we'd use this to do the
discarding later).

An earlier version of this change *never* sent the Message2 with
ReadyForQuery on errors, but that caused errors in simple queries to
make the client hang indefinitely - seemingly there we are expected to
send ReadyForQuery even if it's not in response to a Sync.

There may be more work to be done here in terms of actually discarding
messages until the sync - that's still to be determined - but at least
as of this commit tokio-postgres properly sees errors we're returning in
response to simple query, prepare, and execute messages - all of those
are shown here with a new integration test.

Change-Id: I5411c41fd26e63db8c6bf0029f0b443b869eadc2
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/376
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>