noria-psql: Serialize UnsignedBigInt and Int as INT8
Add support to noria-psql for serializing dataType::UnsignedBigInt and
Datatype::Int as INT8, via directly pattern-matching on both.

A future version of this will likely want to just use
`DataType::coerce_to` to avoid duplicating the logic that's already
there, but for now this gets us one step closer to getting our tests to
pass.

Change-Id: Ia609b7a364ff20213244ed5a97419eef5b0f21e1
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/378
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>