noria-psql: Delete replication slots between tests
It's not sufficient to simply drop the database in between noria-psql
integration tests, as the replicator task still runs in the background
for a little bit after an individual test case is done, and having
active replication slots prevents a database from being dropped.  Rather
than trying to figure out the message-passing that'd be necessary to
properly terminate that replication task in the background, this just
terminates the postgresql backend and deletes the replication slot
before trying to drop the database in between noria-psql integration
tests.

Change-Id: Ib888d8742e51fd02499cc8d2dc09cdb47d0c710c
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/380
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>