noria-psql: Don't try to do fallback reads, for now
Currently due to a deficiency in tokio-postgres, we aren't able to tell
by looking at the response to a postgresql upstream prepare request
whether the query we just prepared was a read or a write, and hence all
UpstreamPrepares returned by the PostgreSqlUpstream are hardcoded to be
writes. This is going to be fixed soon, but in the meantime we can still
get the types tests in particular passing and giving us information
without it, so we can just comment those bits out - this does that, adds
a couple of TODOs for where we need to add things back once we *do* have
proper classification of writes and reads, and adds a new ignored
integration test that'll hopefully be passing once we do.

Change-Id: I3f0594f9b1a24a4215d371acb0a65fafdbda2523
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/381
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>