migrate: Swap out slog for tracing
Swap out all uses of `slog` with `tracing` in
noria/server/src/controller/migrate and all code that depends on it,
which ends up also including DomainHandle, since the slog::Logger field
there was passed in by the migration. This also includes replacing the
construction of *new* loggers with code run in tracing Spans, and one
case (the node for sharding) of adding a new span to factor out a common
field passed to multiple log events. In controller/inner, which is
responsible for actually running migrations and was previously calling
into the entrypoint for migrations, the relevant `tracing` macros are
`use`d local to the function - the idea being that when we switch that
entire module to use tracing that'll give us a warning to remind us to
remove those `use`s.

Refs: ENG-558
Change-Id: Idbea6de3740a714d8196bd6e50230569d53479c8
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/388
Tested-by: Buildkite CI
Reviewed-by: Justin Miron <justin@readyset.io>