Properly manage eviction from MemoryState
Prior to this evictions from MemoryState would fail to properly execute
if a copy of a row was present in the `weak_indices` states. That is
because a `strong_count` would never return a count of 1.
This could lead to an infinite loop in the domain eviction handler.
This change makes sure every evicted key is first removed from `weak_indices`
and only after that it is checked if it was the last reference.

Fixes: ENG-593
Change-Id: Iac71f466f7f7af8baa6a6b2bff18f197ea215e1b
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/387
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>