Replace authority generic with an dispatch enum.
To support dynamically determining authority type at runtime,
this CL removes the authority generic and replaces it with a
dispatch enum. This allows us to create an enum variant when
selecting the authority we are creating. It will be used
in the future to support both zookeeper and consul authorities
in noria-server and noria-mysql/psql.

The enum definition lives in `consensus/mod.rs`. It takes
advantage of the enum_dispatch create to automatically
generate the dispatch calls to each of the variants
LocalAuthority and ZookeeperAuthority.

Change-Id: I23347f6a84365b359e524b9f157feb246f6aee8c
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/392
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>