Parse and ignore precision on DOUBLE
MySQL supports a (deprecated, non-standard) extension to SQL syntax that
allows specifying two precision fields on FLOAT and DOUBLE, to control
the number of digits used when displaying those types to the mysql
shell. We were already parsing that on FLOAT, but weren't on DOUBLE or
DOUBLE PRECISION. This fixes that, and also updates DOUBLE PRECISION to
just be part of the double parse rule since it makes the overall change
shorter.

Change-Id: I924b8108572783710e850d2336742e1d90422bc2
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/394
Tested-by: Buildkite CI
Reviewed-by: Justin Miron <justin@readyset.io>