ops/new-images: Refactor, new base, consul-server
Labeled base as internal base where we also install Tailscale by default
to hopefully move in the direction of having Tailscale installed on
hosts by default instead of using subnet routers.

Started work on a consul-server image which will be experimented with on
deploying with the ops repository as a build cluster for benchmarking.

Added pipeline for creating production internal images similar to the
one created for docker release.

Change-Id: I513e0b03d9d0c62f45ef089a4afe32c4b9a2eaaf
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/395
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>
Reviewed-by: Justin Miron <justin@readyset.io>