data: Blindly coerce from Text to JSON
We need to be able to support the JSON column type since irl.com (a
prospective customer) has a table that uses it - but none of the queries
we want to run actually need to *use* the json, and we're only going to
be deploying with replication enabled, so any validation of the JSON is
going to be done by the upstream db before the data ever gets to us.
Given all of this, it's probably fine to just blindly accept all strings
as JSON for now, and leave the actual validation / parsing to when we
actually want to support JSON (perhaps then we'll even decide we want a
DataType::Json variant).

Change-Id: I2fa286592a0b893bb2a7ae55d177ed0096fb7dd8
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/398
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>