replicators: Read BinlogValue::Jsonb
mysql_common's BinlogValue represents JSON data as its own `Jsonb`
type, separate from the BinlogValue::Value variant (which contains the
actual mysql_common::Value) - to support reading inserted json off the
binlog, we need to match on that variant, convert it to
serde_json::Value, serialize that value as a string, then convert that
string to a DataType.

This does that, and adds a new test to noria-mysql's fallback
integration test suite covering inserting JSON and reading it back.

There's another variant, JsonDiff, which per the mysql documentation[0]
is only ever sent if the binlog_row_value_options system variable is set
to PARTIAL_JSON, which fortunately for us it isn't, since this would be
slightly more complicated for us to support - there's a test covering
partial updates of JSON proving that we support those correctly.

[0]: https://dev.mysql.com/doc/refman/8.0/en/json.html

Change-Id: I74619bcc44482912cd86d9a0d8d893438cd9ca90
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/399
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>