Log all queries when query coverage is setup
This ensures that we log all queries when query coverage is setup. This
is facilitated by creating inner methods for query, prepare, and
execute. By doing so it ensures that no current or future uses of ? to
bubble up errors will cause us to not log an incoming query.

It passes the duty of updating the relevant execute or prepare event to
the inner function, and finalized writing to the outer calling function.

This also adds the following features:
1. Log query duration for both queries hitting ReadySet and upstream.
2. Log any ReadySetError that we get along with the query that generated
   it.
3. Whether fallback was employed.

Remaining TODO for a future CL is to implement a way to compare results
from ReadySet against results from the fallback database, and record
whether the results differ. This is not trivial, and therefore will be
worked on in a future CL.

Change-Id: I6f761193bf9c41d27cb533b9a123ad8b5facea6a
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/402
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>
Reviewed-by: Dan Wilbanks <dan@readyset.io>
Reviewed-by: Michael Cronce <mc@readyset.io>