nom-sql: Add a `test_parse!` macro
This macro encapsulates the common task of running a parser, checking
its output was successful, printing the error *using a string for the
input text following the error*, checking that the remaining input is
empty (also using a string to print it if it isn't empty!) and returning
the parsed result. Going forward I'd like to use this macro for all
parser tests in nom-sql, so that when parsing fails we can actually see
where in the input it failed rather than just getting a raw byte
sequence.

Change-Id: I8aadacab7cbda8dfeddb4063e01a6d914fc2dee2
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/405
Tested-by: Buildkite CI
Reviewed-by: Justin Miron <justin@readyset.io>