nom-sql: Parse index types on [UNIQUE] KEYs
MySQL supports specifying an index type on a KEY definition in a CREATE
TABLE statement, by placing it after the index definition as `USING
HASH` or `USING BTREE`, and IRL.com (a prospective client) uses this in
one of their tables. We should parse this, even if we don't necessarily
use it at all in Noria.

Change-Id: Ie9ce568a863ea2df6fb56fbace58e254b7be249f
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/406
Tested-by: Buildkite CI
Reviewed-by: Justin Miron <justin@readyset.io>